solved the bug of frame id

+++ Vertebra/vertebra/connection/can_receiver.cpp
@@ -5,7 +5,7 @@
 namespace vtb::can
 {
 Receiver::Receiver(
-  Port & port, uint8_t frame_id, std::function<void(const RcvData &)> callback, bool extended)
+  Port & port, uint32_t frame_id, std::function<void(const RcvData &)> callback, bool extended)
 : frame_id_(frame_id), port_(port), callback_(callback)
 {
   if (extended)
@@ -14,7 +14,7 @@ Receiver::Receiver(
     port_.add_std_callback(frame_id_, callback_);
 }
 
-uint8_t Receiver::get_frame_id() { return frame_id_; }
+uint32_t Receiver::get_frame_id() { return frame_id_; }
 }
 
 #endif
